move code from ConfigRepository to ConfigService
